Accidentally pushed a branch that had a longer version of the fix
Revert "Moved version specific Signal.trap implementations to seperate files"

This reverts commit 6d0972f884aa55dd9bc7dbd63e60bd341e2f6ae1.
