+++ kernel/common/load_order18.txt
@@ -85,7 +85,6 @@ random.rbc
 regexp.rbc
 regexp18.rbc
 signal.rbc
-signal18.rbc
 sprinter.rbc
 symbol.rbc
 symbol18.rbc
