+++ kernel/common/load_order19.txt
@@ -97,7 +97,6 @@ random.rbc
 regexp.rbc
 regexp19.rbc
 signal.rbc
-signal19.rbc
 sprinter.rbc
 symbol.rbc
 symbol19.rbc
