+++ kernel/common/load_order20.txt
@@ -93,7 +93,6 @@ random.rbc
 regexp.rbc
 regexp19.rbc
 signal.rbc
-signal19.rbc
 sprinter.rbc
 symbol.rbc
 symbol19.rbc
